Updated Tables
Updated the Spell Table to Auto Increment on the SpellID, the SpellDesc to have more room for Descriptions, the Effect to allow longer character length for Effects.
Update the Database_Builder file to Insert Data and create views on top of it already creating the Database. Altered the format of the Insert statements so that the insertion is one command instead of N. Created the Data Insertion file for the Item Validation Table.